requirements/base.txt
PyQt5==5.12.3
darkdetect
attrs
cattrs
dataset
